ref: extract CreateFormModalContent component from CreateFormModal
