+++ frontend/src/features/workspace/components/CreateFormModal/CreateFormModal.tsx
@@ -1,46 +1,19 @@
-import { FC, useEffect, useState } from 'react'
 import {
-  Box,
-  BoxProps,
   Modal,
   ModalContent,
   useBreakpointValue,
   UseDisclosureReturn,
 } from '@chakra-ui/react'
-import { HTMLMotionProps, motion } from 'framer-motion'
-
-import { Merge } from '~shared/node_modules/type-fest'
 
 import { ModalCloseButton } from '~components/Modal'
 
-import { CreateFormDetailsScreen } from './CreateFormDetailsScreen'
-import {
-  CreateFormFlowStates,
-  CreateFormWizardProvider,
-  useCreateFormWizard,
-} from './CreateFormWizardContext'
-import { SaveSecretKeyScreen } from './SaveSecretKeyScreen'
-
-type MotionBoxProps = Merge<BoxProps, HTMLMotionProps<'div'>>
-const MotionBox: FC<MotionBoxProps> = motion(Box)
+import { CreateModalContent } from './CreateFormModalContent'
+import { CreateFormWizardProvider } from './CreateFormWizardContext'
 
 export type CreateFormModalProps = Pick<
   UseDisclosureReturn,
   'onClose' | 'isOpen'
 >
-
-const SCREEN_ANIMATION_VARIANT = {
-  enter: (direction: number) => {
-    return {
-      x: direction > 0 ? 1000 : -1000,
-      opacity: 0,
-    }
-  },
-  center: {
-    x: 0,
-    opacity: 1,
-  },
-}
 
 export const CreateFormModal = ({
   isOpen,
@@ -58,45 +31,10 @@ export const CreateFormModal = ({
         <ModalCloseButton />
         {isOpen && (
           <CreateFormWizardProvider>
-            <ScreenContent />
+            <CreateModalContent />
           </CreateFormWizardProvider>
         )}
       </ModalContent>
     </Modal>
-  )
-}
-
-/**
- * @preconditions Requires CreateFormWizardProvider parent
- * Display screen content depending on the current step (with animation).
- */
-const ScreenContent = () => {
-  const { direction, currentStep } = useCreateFormWizard()
-  const [isFirstLoad, setIsFirstLoad] = useState(true)
-
-  // So animation does not run on first load.
-  useEffect(() => {
-    if (isFirstLoad) {
-      setIsFirstLoad(false)
-    }
-  }, [isFirstLoad])
-
-  return (
-    <MotionBox
-      key={currentStep}
-      custom={direction}
-      variants={SCREEN_ANIMATION_VARIANT}
-      initial={isFirstLoad ? 'center' : 'enter'}
-      animate="center"
-      transition={{
-        x: { type: 'spring', stiffness: 300, damping: 30 },
-        opacity: { duration: 0.2 },
-      }}
-    >
-      {currentStep === CreateFormFlowStates.Details && (
-        <CreateFormDetailsScreen />
-      )}
-      {currentStep === CreateFormFlowStates.Landing && <SaveSecretKeyScreen />}
-    </MotionBox>
   )
 }
